+++ src/app/not-found.tsx
@@ -1,14 +1,10 @@
-import { Header } from '@/src/components/common/header';
-import { Footer } from '@/src/components/common/footer';
+import { WrappedScreenContentLayout } from '@/src/components/common/layouts/wrapped-screen-content-layout';
 import Link from 'next/link';
-import styles from '@/styles/app/not-found.module.scss';
 
 export default function NotFound() {
     return (
-        <div className={styles['not-found']}>
-            <Header />
-            <div className={styles['not-found__info']}>This page does not exist. Go to&nbsp;<Link href="/">main</Link>&nbsp;page.</div>
-            <Footer />
-        </div>
+        <WrappedScreenContentLayout>
+            This page does not exist. Go to&nbsp;<Link href="/">main</Link>&nbsp;page.
+        </WrappedScreenContentLayout>
     );
 }